requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.1
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-tinymce==3.0.2
Pillow==7.2.0
pycodestyle==2.6.0
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
